move components out of routing
