+++ src/components/ContentPlaceholder.tsx
