+++ src/components/FileError.tsx
