+++ src/components/RegionBanner.tsx
