+++ src/views/submission/clinical-submission/components/FilesNavigator/ui.tsx
@@ -18,10 +18,10 @@
  */
 
 import { ContentPlaceholder } from '@/app/components/ContentPlaceholder';
-import ErrorNotification from '@/app/components/ErrorNotification';
+import ErrorNotification from '@/components/ErrorNotification';
 import ErrorNotificationDefaultTable, {
 	getDefaultErrorReportColumns,
-} from '@/app/components/ErrorNotification/ErrorNotificationDefaultTable';
+} from '@/components/ErrorNotification/ErrorNotificationDefaultTable';
 import { toDisplayError } from '@/global/utils';
 import { css } from '@/lib/emotion';
 import { NOTIFICATION_VARIANTS } from '@icgc-argo/uikit';
