+++ src/views/submission/dashboard/components/DonorDataSummary/DonorSummaryTable.tsx
@@ -18,10 +18,10 @@
  */
 
 import { DonorSummaryEntry } from '@/__generated__/gateway/graphql';
-import { CellContentCenter, DataTableStarIcon as StarIcon } from '@/app/components/Table/common';
-import { ContentError } from '@/app/components/placeholders';
 import CLINICAL_ERRORS_QUERY from '@/app/gql/clinical/CLINICAL_ERRORS_QUERY';
 import { useClinicalQuery } from '@/app/hooks';
+import { CellContentCenter, DataTableStarIcon as StarIcon } from '@/components/Table/common';
+import { ContentError } from '@/components/placeholders';
 import { displayDate, parseDonorIdString } from '@/global/utils';
 import { css } from '@/lib/emotion';
 import {
